extract reducerWithState to a separate file
